rembg/requirements.txt
flask==3.0.0
rembg[cpu]==2.0.57
pillow>=10.0.0
gunicorn==21.2.0
